training: move metric design into a specific module
